범위 내에 몬스터 접근시 공격 구현

+++ Guardner/Assets/OtherScript/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Ondamage(float damage, Vector2 hitPoint, Vector2 hitNormal);
+    void Ondamage(int damage, Vector2 hitPoint, Vector2 hitNormal);
 
     void Die();
 }
